requirements.txt
streamlit
pandas
cloudinary
Pillow
requests